Fix GPS after resume
On Thu, 22 Nov 2012, Torsten Duwe wrote:

> It's the software equivalent of soldering/clamping RX and TX together.
[...]
> I'll continue to investigate next week...

Unfortunately the loopback happens so far inside the chip that the
electrical I/O and the special function config circuit is not used.
Some code during boot configures GPIO H5 as RxD1. The deep standby resets
it, and it wakes up in its default mode: a simple input.
Fix attached.